+++ R/data_sufficiency.R
@@ -33,8 +33,8 @@
 #' @param midf_table `r midfield_x("term")` with required variables
 #'        `{mcid, term, institution}.`
 #' @returns Data frame with the following properties:
-#' * `r preserv_class_not_grp_keys`
-#' * `r omit_NA_dup_rows`
+#' * `r keep_class_rm_groups_rm_keys`
+#' * `r keep_row_order_rm_NA_rm_duplic_rows`
 #' * `r add_new_cols` The new variables are:
 #'   - `data_range` &nbsp; Character. Institution data range, encoded
 #'     `YYYYT-YYYYT,` indicating the institution's first and last term in the